move DisciplineFilter component into filters folder #243
